+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## v0.5.5
 - 10/27/2024
+- Integrated Scalar openapi docs with fastapi
+- Added Image preview for assets page
 - Complete Asset CRUD page
 
 ## v0.5.2